# ARROW-15056: [C++] Speed up GcsFileSystem tests