This reduces the running time from about 60s to about 10s.

Before:

```
86/86 Test #50: arrow-gcsfs-test ..........................   Passed   60.00 sec
```

After:

```
71/86 Test #50: arrow-gcsfs-test ..........................   Passed    9.62 sec
```