@pitrou 2021-12-13
Wow, that's nifty, thank you!